userpro/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-model-utils==4.3.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
